[irfan] now playing movies home + commons utils + animation fade

+++ app/src/main/java/id/co/moviedb/data/MoviesResponse.kt
@@ -26,6 +26,6 @@ data class MoviesModel(
     @SerializedName("release_date") val releaseDate: String?,
     @SerializedName("title") val title: String?,
     @SerializedName("video") val video: Boolean?,
-    @SerializedName("vote_average") val voteAverage: Int?,
+    @SerializedName("vote_average") val voteAverage: Float?,
     @SerializedName("vote_count") val voteCount: Int?
 )
